ai/requirements.txt
requests
python-dotenv
pydantic-ai
ollama